secure RedisService

+++ src/services/RedisService.js
@@ -1,5 +1,6 @@
 const redis = require('redis');
 const { REDIS_URI } = require('../constants/Constants');
+const { missingKey, missingValue } = require('../constants/serviceErrors');
 
 const client = redis.createClient(REDIS_URI);
 
@@ -8,18 +9,34 @@ const client = redis.createClient(REDIS_URI);
 })();
 
 const get = async (key) => {
-  const value = await client.get(key);
-  return value;
+  if (!key) throw missingKey;
+  try {
+    const value = await client.get(key);
+    return value;
+  } catch (err) {
+    return err;
+  }
 };
 
 const set = async (key, value) => {
-  const response = await client.setEx(key, 864000, value);
-  return response;
+  if (!key) throw missingKey;
+  if (!value) throw missingValue;
+  try {
+    const response = await client.setEx(key, 864000, value);
+    return response;
+  } catch (err) {
+    return err;
+  }
 };
 
 const del = async (key) => {
-  const response = await client.del(key);
-  return response;
+  if (!key) throw missingKey;
+  try {
+    const response = await client.del(key);
+    return response;
+  } catch (err) {
+    return err;
+  }
 };
 
 const disconnect = () => {
